# api/openapi/registration.yaml
openapi: "3.0.0"
info:
  title: "Registration & Onboarding API"
  version: "1.0.0"
  description: |
    Bot-facing endpoints covering the pilot registration and VA onboarding flow.
    Consumed by comrade-bot to register pilots, initialise Discord servers as VAs,
    join memberships, and look up user status.

    All endpoints require authentication via three headers sent by the bot.
servers:
  - url: /api/v1
security:
  - ApiKey: []
    ServerId: []
    DiscordId: []
paths:
  /pilots/register:
    post:
      operationId: registerPilot
      summary: Register a pilot with IFC credentials
      description: |
        Links a Discord account to an Infinite Flight Community (IFC) account.
        Verifies the account exists and the supplied last-flight route matches
        the IFC logbook. Returns a flag indicating whether the server is already
        set up as a VA (so the bot knows whether to prompt the user to /join).
      tags: [registration]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterPilotRequest"
      responses:
        "201":
          description: Pilot registered successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RegisterPilotResponse"
        "400":
          description: Missing or invalid fields, or last-flight mismatch
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Missing or invalid X-API-Key authentication
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Missing required Discord context headers X-Discord-User-Id or X-Discord-Server-Id
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: User or IFC ID already registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "422":
          description: Validation failure (missing required fields)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ValidationErrorResponse"
        "500":
          description: Upstream IFC API or database error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /server/init:
    post:
      operationId: initServer
      summary: Bootstrap a Discord server with a VA Code / ID
      description: |
        Creates a minimal VA record linked to the Discord server in X-Discord-Server-Id.
        Requires the calling user to already be registered (/pilots/register) and
        makes that user the VA admin. Discord collects only the VA Code / ID;
        flight matching setup continues in Vizburo.
      tags: [registration]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InitServerRequest"
      responses:
        "201":
          description: Server bootstrapped and minimal VA created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InitServerResponse"
        "400":
          description: Missing fields or unregistered admin prerequisite
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Missing or invalid X-API-Key authentication
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Missing required Discord context headers X-Discord-User-Id or X-Discord-Server-Id
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Discord server already registered as a VA, or VA Code / ID already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "422":
          description: Validation failure (missing required fields)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ValidationErrorResponse"
        "500":
          description: Database error during VA creation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /memberships/join:
    post:
      operationId: joinMembership
      summary: Join the VA associated with the current Discord server
      description: |
        Creates a membership record linking the authenticated Discord user to the VA
        resolved from X-Discord-Server-Id. The callsign must exist in the VA's linked Airtable
        base; the bot should surface the Airtable URL in the error message when it
        does not.
      tags: [registration]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/JoinMembershipRequest"
      responses:
        "201":
          description: Membership created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JoinMembershipResponse"
        "400":
          description: Missing callsign, invalid format, or callsign not found in Airtable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Missing or invalid X-API-Key authentication
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Missing required Discord context headers X-Discord-User-Id or X-Discord-Server-Id
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: VA not found for this Discord server
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: User is already a member of this VA, or callsign is taken
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "422":
          description: Validation failure (missing required fields)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ValidationErrorResponse"
        "500":
          description: Unexpected database error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /user/status:
    get:
      operationId: getUserStatus
      summary: Get authenticated user's registration and membership status
      description: |
        Returns explicit onboarding state for the authenticated Discord user and
        current Discord server. Normal onboarding states, including unregistered
        users and servers that are not configured as VAs, are represented as a
        200 response rather than requiring clients to infer state from 404s.
      tags: [registration]
      responses:
        "200":
          description: User status retrieved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserStatusResponse"
        "401":
          description: Missing or invalid X-API-Key authentication
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Missing required Discord context headers X-Discord-User-Id or X-Discord-Server-Id
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Database error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /signed-link:
    post:
      operationId: generateSignedLink
      summary: Generate a time-limited signed URL for dashboard access
      description: |
        Produces a presigned URL that, when visited, creates a browser session
        for the authenticated Discord user without a password. The bot posts this
        link so the user can access the Vizburo dashboard from Discord.
        TTL defaults to 15 minutes if omitted; redirect_to defaults to /dashboard.
      tags: [registration]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GenerateSignedLinkRequest"
      responses:
        "200":
          description: Signed link generated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenerateSignedLinkResponse"
        "400":
          description: Malformed request body
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Missing or invalid X-API-Key authentication
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Missing required Discord context headers X-Discord-User-Id or X-Discord-Server-Id
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: User or VA not found for the given Discord IDs
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Token generation failure
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  securitySchemes:
    ApiKey:
      type: apiKey
      in: header
      name: X-API-Key
      description: Bot API key issued per VA
    ServerId:
      type: apiKey
      in: header
      name: X-Discord-Server-Id
      description: Discord server (guild) ID — used to resolve the VA context
    DiscordId:
      type: apiKey
      in: header
      name: X-Discord-User-Id
      description: Discord user ID of the caller
  schemas:
...
    ErrorResponse:
      type: object
      required: [status, responseTimeMs]
      properties:
        status:
          type: string
          enum: [error]
        error:
          $ref: "#/components/schemas/ErrorDetail"
        responseTimeMs:
          type: integer
          format: int64
    ValidationErrorResponse:
      type: object
      required: [status, responseTimeMs]
      properties:
        status:
          type: string
          enum: [error]
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              enum: [VALIDATION_FAILED]
            message:
              type: string
            fields:
              type: array
              items:
                $ref: "#/components/schemas/ValidationFieldError"
        responseTimeMs:
          type: integer
          format: int64

    # ── /pilots/register ────────────────────────────────────────────────────
    RegisterPilotRequest:
      type: object
      required: [ifc_id, last_flight]
      properties:
        ifc_id:
          type: string
          description: Infinite Flight Community username
          example: john_doe
        last_flight:
          type: string
          description: |
            IATA or ICAO route of the pilot's most recent Infinite Flight flight,
            used to verify ownership of the IFC account (e.g. "KLAX-KSFO").
          example: KLAX-KSFO
...
    RegisterPilotResponse:
      type: object
      required: [status, responseTimeMs]
      properties:
        status:
          type: string
          enum: [ok]
        result:
          $ref: "#/components/schemas/RegisterPilotData"
        responseTimeMs:
          type: integer
          format: int64

    # ── /server/init ────────────────────────────────────────────────────────
    InitServerRequest:
      type: object
      required: [va_code]
      properties:
        va_code:
          type: string
          description: Short admin-facing VA Code / ID (e.g. "IFE")
          example: IFE
...
    InitServerResponse:
      type: object
      required: [status, responseTimeMs]
      properties:
        status:
          type: string
          enum: [ok]
        result:
          $ref: "#/components/schemas/InitServerData"
        responseTimeMs:
          type: integer
          format: int64

    # ── /memberships/join ───────────────────────────────────────────────────
    JoinMembershipRequest:
      type: object
      required: [callsign]
      properties:
        callsign:
          type: string
          description: Pilot's callsign as it appears in the VA's Airtable base
          example: "123"
...
    JoinMembershipResponse:
      type: object
      required: [status, responseTimeMs]
      properties:
        status:
          type: string
          enum: [ok]
        result:
          $ref: "#/components/schemas/JoinMembershipData"
        responseTimeMs:
          type: integer
          format: int64

    # ── /user/status ────────────────────────────────────────────────────────
...
    UserStatusResponse:
      type: object
      required: [status, responseTimeMs]
      properties:
        status:
          type: string
          enum: [ok]
        result:
          $ref: "#/components/schemas/UserStatusData"
        responseTimeMs:
          type: integer
          format: int64

    # ── /signed-link ────────────────────────────────────────────────────────
    GenerateSignedLinkRequest:
      type: object
      properties:
        redirect_to:
          type: string
          description: Path to redirect the user to after login. Defaults to /dashboard.
          example: /dashboard
        ttl_minutes:
          type: integer
          description: Link validity in minutes. Defaults to 15.
          minimum: 1
          maximum: 1440
          example: 15
...
    GenerateSignedLinkResponse:
      type: object
      required: [status, responseTimeMs]
      properties:
        status:
          type: string
          enum: [ok]
        result:
          $ref: "#/components/schemas/GenerateSignedLinkData"
        responseTimeMs:
          type: integer
          format: int64